"Remove unused psalm.xml, update PHPUnit exception handling"
Removal of psalm.xml was done as this file is not used anywhere in the project. Changes in phpstan.dist.neon, composer.json, BasicAuthTest.php, TestCase.php and FixturesKernel.php are for updating PHPUnit tests to handle 'UnexpectedAlertOpenException' in a more efficient way and updating packages. In composer.json, 'php-webdriver/webdriver' package is added into the required-dev dependency list for PHPUnit testing. In TestCase.php, unnecessary methods for setting up and tearing down are removed to simplify the code and increase readability. In FixturesKernel.php, a more descriptive variable name is used for clarity.